requirements.txt
torch>=2.2.2,<2.6; sys_platform == 'darwin' and platform_machine == 'x86_64'  # Intel macOS
torch >= 2.2.2; sys_platform != 'darwin' or platform_machine != 'x86_64'   # all chapters
jupyterlab >= 4.0          # all
tiktoken >= 0.5.1          
matplotlib >= 3.7.1        
tensorflow>=2.16.2; sys_platform == 'darwin' and platform_machine == 'x86_64'  # Intel macOS
tensorflow >= 2.18.0; sys_platform != 'darwin' or platform_machine != 'x86_64'   
tqdm >= 4.66.1             
numpy >= 1.26, < 2.1       # dependency of several other libraries like torch and pandas
pandas >= 2.2.1            # for experiment comparison and analysis
psutil >= 5.9.5            # already installed automatically as dependency of torch
pyyaml >= 6.0              # for experiment configuration management
lightning >= 2.0.0         # for PyTorch Lightning framework
lightning_sdk >= 2025.0.0  # for Lightning AI Studio cloud storage and management

# OpenAI API for model evaluation and testing
openai >= 1.0.0              # for OpenAI API integration
requests >= 2.31.0           # HTTP library for API calls

# Optional dependencies for RAG functionality (uncomment if needed)
# faiss-cpu >= 1.7.4       # for RAG retrieval (CPU version)
# OR
# faiss-gpu >= 1.7.4       # for RAG retrieval (GPU version)
